% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_tuning.R
\name{plot.spaDesign2Tuning}
\alias{plot.spaDesign2Tuning}
\title{Plot Method for spaDesign2Tuning}
\usage{
\method{plot}{spaDesign2Tuning}(x, ...)
}
\arguments{
\item{x}{An object of class \code{spaDesign2Tuning}.}

\item{...}{Additional arguments to be passed to plotting functions.}
}
\description{
Plot Method for spaDesign2Tuning
}
