% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_tuning.R
\name{summary.spaDesign2Tuning}
\alias{summary.spaDesign2Tuning}
\title{Summary Method for spaDesign2Tuning}
\usage{
\method{summary}{spaDesign2Tuning}(object, ...)
}
\arguments{
\item{object}{An object of class \code{spaDesign2Tuning}.}

\item{...}{Additional arguments affecting the summary produced.}
}
\description{
Summary Method for spaDesign2Tuning
}
