% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSpaDesign2Object.R
\name{syntheticData}
\alias{syntheticData}
\alias{syntheticData,spaDesign2-method}
\alias{syntheticData<-}
\alias{syntheticData<-,spaDesign2-method}
\title{Get or Set Synthetic Data}
\usage{
syntheticData(object)

\S4method{syntheticData}{spaDesign2}(object)

syntheticData(object) <- value

\S4method{syntheticData}{spaDesign2}(object) <- value
}
\arguments{
\item{object}{A \code{spaDesign2} object.}

\item{value}{A named list of simulated samples to store.}
}
\value{
\code{syntheticData} returns the stored list; the replacement method
returns the updated \code{spaDesign2} object.
}
\description{
Accessor and replacement methods for the \code{synthetic_data}
slot of a \code{spaDesign2} object.
}
